Release System PowerUp

+++ Assets/Scripts/HUD/PowerUpCards.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu( fileName = "powerUpCards",menuName = "PowerUp Card/New Card")]
@@ -8,7 +6,7 @@ public class PowerUpCards : ScriptableObject
     public int ID;
     public Sprite spriteRender;
     public int bibleTotal;
-    public int hadoukenTime;
+    public float hadoukenTime;
     public float darkPowerArea;
     public float timeShield;
     public float ShieldInterval;
